Add generic contract class providing common functionality for any specific contract client
